search functions moved from home to searchForm component
