+++ client/src/Components/SearchForm/SearchForm.jsx
@@ -1,10 +1,33 @@
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
+import { useState } from "react";
 import './SearchForm.css';
 
-const SearchForm = ({ handleSearch, handleChange, searchTerm }) => {
+const SearchForm = ({ handleSearch }) => {
+
+  console.log("The imported function is:", handleSearch)
+
+  const [searchTerm, setSearchTerm] = useState("");
+
+  const handleChange = (evt) => {
+    evt.preventDefault();
+
+    const term = evt.target.value;
+    console.log("The new search term is:", term);
+
+    setSearchTerm((currTerm) => {
+      currTerm = term;
+      return currTerm;
+    });
+  };
+  
+  const handleSubmit = (evt) => {
+    evt.preventDefault();
+    handleSearch(searchTerm);
+  };
+  
   return (
-    <Form className="d-flex" onSubmit={handleSearch} id="search-form">
+    <Form className="d-flex" onSubmit={handleSubmit} id="search-form">
       <Form.Control
         type="search"
         placeholder="Search for the title or author"
