+++ client/src/Pages/Home/Home.css
@@ -12,3 +12,7 @@
     align-items: center;
     margin-top: 50px;
   }
+
+  .error {
+    color: red;
+  }
